Altered the tests to be performed in a separate class
