+++ src/test/java/com/sparta/spartaSimulator/AppTest.java
@@ -12,16 +12,4 @@ class AppTest {
         Assertions.assertEquals(1, 1);
     }
 
-    @Test
-    public void getSetCurrentTimeTest() {
-        TimeManager.setCurrentTime(3);
-        Assertions.assertEquals(3, TimeManager.getCurrentTime());
-    }
-
-    @Test
-    public void delayTimeTest() {
-        TimeManager.delayTime(2, 4);
-        Assertions.assertEquals(8, TimeManager.getCurrentTime());
-    }
-
 }
